+++ docs/registry/registry-ui.ts
@@ -335,7 +335,7 @@ export const registryUI: Registry = {
         },
         {
           path: "list-header.tsx",
-          dependencies: { "@seed-design/react": "~0.3.0", "@seed-design/css": "~0.3.0" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },