+++ packages/figma/src/entities/data/__generated__/styles/index.mjs
@@ -1,0 +1,296 @@
+export const FIGMA_STYLES = [
+  {
+    "styleType": "FILL",
+    "key": "c1cee6b89c357c6a12d020654d73a4490cafdf51",
+    "name": "gradient/fade/layer-default/←(to-left)",
+    "description": "스크롤이 가능하도록 그라디언트로 특정 영역을 덮는 요소로 사용해요.",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "ff436e73cf9477a1006e6f8a1f9d2c9dbe0e5edd",
+    "name": "gradient/fade/layer-default/→(to-right)",
+    "description": "스크롤이 가능하도록 그라디언트로 특정 영역을 덮는 요소로 사용해요.",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "826457503b7ee3d472eab91581b1fcd499c47c00",
+    "name": "gradient/fade/layer-default/↑(to-top)",
+    "description": "스크롤이 가능하도록 그라디언트로 특정 영역을 덮는 요소로 사용해요.",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "408cb970f1de672217f452152b0398a4bb4f4e12",
+    "name": "gradient/fade/layer-default/↓(to-bottom)",
+    "description": "스크롤이 가능하도록 그라디언트로 특정 영역을 덮는 요소로 사용해요.",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "df22450a8f8e9c323baf40d8981920d1b6376cbf",
+    "name": "gradient/fade/layer-floating/←(to-left)",
+    "description": "스크롤이 가능하도록 그라디언트로 특정 영역을 덮는 요소로 사용해요.",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "063c0dc05b438ac36499492e783520a722782d74",
+    "name": "gradient/fade/layer-floating/→(to-right)",
+    "description": "스크롤이 가능하도록 그라디언트로 특정 영역을 덮는 요소로 사용해요.",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "46186d7dd037e8e54983ee6677cf7915790d84cc",
+    "name": "gradient/fade/layer-floating/↑(to-top)",
+    "description": "스크롤이 가능하도록 그라디언트로 특정 영역을 덮는 요소로 사용해요.",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "b28d9c9adc816fdd4ecec3a0b3e25d8aec802b6a",
+    "name": "gradient/fade/layer-floating/↓(to-bottom)",
+    "description": "스크롤이 가능하도록 그라디언트로 특정 영역을 덮는 요소로 사용해요.",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "353baa82e0440385eb121cd0ae98d2fcb1a70b18",
+    "name": "gradient/glow/magic",
+    "description": "반짝이는 것처럼 느껴지는 배경에 쓰이는 ai 컬러입니다.",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "f83f54c1840aaf26742629e2c696c363b8df0a82",
+    "name": "gradient/glow/magic-pressed",
+    "description": "반짝이는 것처럼 느껴지는 배경에 쓰이는 ai 컬러의 pressed컬러입니다.",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "1101a7c6a505d97904c3a0ce6a9ee17fe15de879",
+    "name": "gradient/highlight/magic",
+    "description": "아이콘 및 shape 영역에서 AI 기능을 표현할 때 사용하는 컬러입니다.",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "528a3808331ad87aacb47430a59cd32d4410b822",
+    "name": "gradient/highlight/magic-pressed",
+    "description": "",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "2ed61bff1a187f39389698cf63b74d6997aed609",
+    "name": "gradient/shimmer/magic",
+    "description": "Skeleton AI shimmer",
+    "remote": false
+  },
+  {
+    "styleType": "FILL",
+    "key": "71971dde923a11c615296a16f8172c0671d5a3a6",
+    "name": "gradient/shimmer/neutral",
+    "description": "Skeleton shimmer",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "af24f9e7cc90af3bf2a18029dd59ae0646966486",
+    "name": "legacy-13-150",
+    "description": "13",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "2fec4e47f53f75ea46bee4d740795e099af5472f",
+    "name": "scale/legacy-14-150",
+    "description": "14",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "337e113e2fafc22ccd5423b3fc7a7da69a6c917c",
+    "name": "scale/t1-bold",
+    "description": "11",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "49735ed4693e2f7e9d98b6b6e3818f38e2fe77a8",
+    "name": "scale/t1-medium",
+    "description": "11",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "4d54fb58fa4a63a6ddec973099b4babfbe502849",
+    "name": "scale/t1-regular",
+    "description": "11",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "6ea6b06312771259d59de265a5ef12cd1dae9102",
+    "name": "scale/t10-bold",
+    "description": "26",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "8378bd59b14f5682dcaf20110efde4b7a1fb18e3",
+    "name": "scale/t2-bold",
+    "description": "12",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "2aff06b0ad8be136acbfb3090fa4dde1f9eac48e",
+    "name": "scale/t2-medium",
+    "description": "12",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "711eaf0bd91b906e108f3508cf7845fc82294b1a",
+    "name": "scale/t2-regular",
+    "description": "12",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "f749d659a689cbbecbdbb1b559056731234332a3",
+    "name": "scale/t3-bold",
+    "description": "13",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "96a2f248d201c246587d63eb4d00cb7610897400",
+    "name": "scale/t3-medium",
+    "description": "13",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "22ead16ee07b342df4a26952bc987f3155988779",
+    "name": "scale/t3-regular",
+    "description": "13",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "a85ea49c1625319427c047bc206c26e708012030",
+    "name": "scale/t4-bold",
+    "description": "14",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "2c143ba862b3c1af10a0bd5d2e723bd12e7878ef",
+    "name": "scale/t4-medium",
+    "description": "14",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "b23d3e206b0244690c19472fd78bfc33b9b46d06",
+    "name": "scale/t4-regular",
+    "description": "14",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "8eff229c53f06eeb418f39ad1cb3dbe12480f12b",
+    "name": "scale/t5-bold",
+    "description": "16",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "bc8bcf3caa904cdc2e1c2983d0f12d32832d7374",
+    "name": "scale/t5-medium",
+    "description": "16",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "54a5a4c2bfc09573fba3a8efe0dafff067cf53bd",
+    "name": "scale/t5-regular",
+    "description": "16",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "adf921a56e89b5737a6b626034cfea184e828370",
+    "name": "scale/t6-bold",
+    "description": "18",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "02ee25a9050a9c78dbaa03ea006aa4d1db593e99",
+    "name": "scale/t6-medium",
+    "description": "18",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "6608eaab965859080a5effa6e23eb857d888057f",
+    "name": "scale/t6-regular",
+    "description": "18",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "182d9252d7f7380075915a89e6160d7595124bc3",
+    "name": "scale/t7-bold",
+    "description": "20",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "b2e24640e7dc1264fc638434511c9f994e1e149c",
+    "name": "scale/t7-medium",
+    "description": "20",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "eb1c0409582609a51fe9afc9e31301d529e3d12f",
+    "name": "scale/t7-regular",
+    "description": "20",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "b7f864e3f64524a863d38729e2cea18f44c7ebaa",
+    "name": "scale/t8-bold",
+    "description": "22",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "b146e1317c67db787834f1890493225bdbba4e48",
+    "name": "scale/t9-bold",
+    "description": "24",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "85422ef9f0230f821f1a9e6d2f8f86b3ba87f20d",
+    "name": "semantic/article-body",
+    "description": "16",
+    "remote": false
+  },
+  {
+    "styleType": "TEXT",
+    "key": "d918eb7069e10c8dd745abf40bc375004c1badb8",
+    "name": "semantic/screen-title",
+    "description": "26",
+    "remote": false
+  }
+];