+++ packages/react-headless/popover/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @seed-design/react-popover
 
+## 1.0.1
+
+### Patch Changes
+
+- b10ff0b: closeOnInteractOutside를 false로 설정하여 Help Bubble 외부와 상호작용 시에도 닫히지 않도록 설정할 수 있습니다. (기본값: true)
+
 ## 1.0.0
 
 ### Major Changes