+++ packages/react/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @seed-design/react
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [b10ff0b]
+  - @seed-design/react-popover@1.0.1
+
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [0b8a02e]
+- Updated dependencies [6c6099d]
+  - @seed-design/css@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [6d2e13d]
+  - @seed-design/css@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes