+++ packages/rootage/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @seed-design/rootage-artifacts
 
+## 1.0.3
+
+### Patch Changes
+
+- 6c6099d: Callout에 tone=positive variant를 추가합니다.
+
+## 1.0.2
+
+### Patch Changes
+
+- 6d2e13d: MannerTemp 컴포넌트가 레이아웃에서 너비를 덜 차지하도록 업데이트합니다.
+
 ## 1.0.1
 
 ### Patch Changes